# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface DeleteResponse {
  success: boolean;
  message: string;
}

export interface BulkDeleteRequest {
  ids: number[];
}

export interface Purchase {
  id: number;
  nomor: string;
  tanggal: string;
  keterangan: string;
  jumlah: number;
  satuan: string;
  harga_satuan: number;
  harga_total: number;
  catatan: string;
  kategori: string;
  supplier?: string | null;
  supplier_contact?: string | null;
}

export interface PurchaseInput {
  nomor: string;
  tanggal: string;
  keterangan: string;
  jumlah: number;
  satuan: string;
  harga_satuan: number;
  catatan: string;
  kategori: string;
  supplier?: string | null;
  supplier_contact?: string | null;
}

export interface PurchasePlan {
  id: number;
  nomor: string;
  tanggal: string;
  keterangan: string;
  jumlah: number;
  satuan: string;
  harga_satuan: number;
  harga_total: number;
  catatan: string;
  kategori: string;
  supplier?: string | null;
  supplier_contact?: string | null;
}

export interface PurchasePlanInput {
  nomor: string;
  tanggal: string;
  keterangan: string;
  jumlah: number;
  satuan: string;
  harga_satuan: number;
  catatan: string;
  kategori: string;
  supplier?: string | null;
  supplier_contact?: string | null;
}

export interface DashboardSummary {
  total_purchases: number;
  total_purchase_plans: number;
  total_spend: number;
  total_planned: number;
  this_month_purchases: number;
  this_month_plans: number;
}

export interface RecentActivity {
  recent_purchases: Purchase[];
  recent_plans: PurchasePlan[];
}

export interface BackupResponse {
  success: boolean;
  filename: string;
  message: string;
}

export interface BackupFile {
  filename: string;
  size: number;
  created: string;
}

export type GetPurchasesParams = {
  search?: string;
  startDate?: string;
  endDate?: string;
  kategori?: string;
};

export type ExportPurchasesExcelParams = {
  search?: string;
  startDate?: string;
  endDate?: string;
  kategori?: string;
  /**
   * Comma-separated IDs for selective export
   */
  ids?: string;
};

export type ExportPurchasesPdfParams = {
  search?: string;
  startDate?: string;
  endDate?: string;
  kategori?: string;
  /**
   * Comma-separated IDs for selective export
   */
  ids?: string;
};

export type GetPurchasePlansParams = {
  search?: string;
  startDate?: string;
  endDate?: string;
  kategori?: string;
};

export type ExportPurchasePlansExcelParams = {
  search?: string;
  startDate?: string;
  endDate?: string;
  kategori?: string;
  /**
   * Comma-separated IDs for selective export
   */
  ids?: string;
};

export type ExportPurchasePlansPdfParams = {
  search?: string;
  startDate?: string;
  endDate?: string;
  kategori?: string;
  /**
   * Comma-separated IDs for selective export
   */
  ids?: string;
};