# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /purchase-plans:
    get:
      operationId: getPurchasePlans
      tags: [purchase-plans]
      summary: Get all purchase plans
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: startDate
          in: query
          required: false
          schema:
            type: string
        - name: endDate
          in: query
          required: false
          schema:
            type: string
        - name: kategori
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of purchase plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchasePlan"
    post:
      operationId: createPurchasePlan
      tags: [purchase-plans]
      summary: Create a new purchase plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchasePlanInput"
      responses:
        "201":
          description: Created purchase plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchasePlan"
  /purchase-plans/{id}:
    get:
      operationId: getPurchasePlan
      tags: [purchase-plans]
      summary: Get a purchase plan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchasePlan"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updatePurchasePlan
      tags: [purchase-plans]
      summary: Update a purchase plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchasePlanInput"
      responses:
        "200":
          description: Updated purchase plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchasePlan"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePurchasePlan
      tags: [purchase-plans]
      summary: Delete a purchase plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /purchase-plans/bulk-delete:
    post:
      operationId: bulkDeletePurchasePlans
      tags: [purchase-plans]
      summary: Delete multiple purchase plans by IDs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkDeleteRequest"
      responses:
        "200":
          description: Bulk delete result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
  /purchase-plans/export/excel:
    get:
      operationId: exportPurchasePlansExcel
      tags: [purchase-plans]
      summary: Export purchase plans as Excel file
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: startDate
          in: query
          required: false
          schema:
            type: string
        - name: endDate
          in: query
          required: false
          schema:
            type: string
        - name: kategori
          in: query
          required: false
          schema:
            type: string
        - name: ids
          in: query
          required: false
          schema:
            type: string
          description: Comma-separated IDs for selective export
      responses:
        "200":
          description: Excel file download
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
  /purchase-plans/export/pdf:
    get:
      operationId: exportPurchasePlansPdf
      tags: [purchase-plans]
      summary: Export purchase plans as PDF file
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: startDate
          in: query
          required: false
          schema:
            type: string
        - name: endDate
          in: query
          required: false
          schema:
            type: string
        - name: kategori
          in: query
          required: false
          schema:
            type: string
        - name: ids
          in: query
          required: false
          schema:
            type: string
          description: Comma-separated IDs for selective export
      responses:
        "200":
          description: PDF file download
          content:
            application/pdf:
              schema:
                type: string
                format: binary
...
  /backup/download/{filename}:
    get:
      operationId: downloadBackup
      tags: [backup]
      summary: Download a backup file
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Backup file download
          content:
            application/sql:
              schema:
                type: string
                format: binary
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    BulkDeleteRequest:
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
      required:
        - ids
...
    PurchasePlan:
      type: object
      properties:
        id:
          type: integer
        nomor:
          type: string
        tanggal:
          type: string
        keterangan:
          type: string
        jumlah:
          type: number
        satuan:
          type: string
        harga_satuan:
          type: number
        harga_total:
          type: number
        catatan:
          type: string
        kategori:
          type: string
        supplier:
          type: string
          nullable: true
        supplier_contact:
          type: string
          nullable: true
      required:
        - id
        - nomor
        - tanggal
        - keterangan
        - jumlah
        - satuan
        - harga_satuan
        - harga_total
        - catatan
        - kategori
    PurchasePlanInput:
      type: object
      properties:
        nomor:
          type: string
        tanggal:
          type: string
        keterangan:
          type: string
        jumlah:
          type: number
        satuan:
          type: string
        harga_satuan:
          type: number
        catatan:
          type: string
        kategori:
          type: string
        supplier:
          type: string
          nullable: true
        supplier_contact:
          type: string
          nullable: true
      required:
        - nomor
        - tanggal
        - keterangan
        - jumlah
        - satuan
        - harga_satuan
        - catatan
        - kategori
...